rwp-function-app/requirements.txt
azure-functions>=1.17.0
azure-identity>=1.15.0
aioodbc>=0.5.0
